Advanced post editor support. Emoticon popup menu.
This is the cool update. First, emoticons work with the advanced post
editor (Go Advanced, editing a post, and posting a new thread), but only
in the WYSIWYG-off mode. Second, the vanilla emoticon popup menu is
replaced with a cool new one that includes both Facepunch and Something
Awful emoticons. SA Emoticons have a search box that works with magical fuzzy search and it's all magical.